Restrict sending to user ID
Figured out a possible way to have the pictures that are sent be restricted to people in the game using the User class. When a picture is taken, the User class' ID gets updated  along with the picture.

Then, I think, it will work by having that picture linked with the user ID who took the picture.

This allows us to then send the picture using the user ID who took the picture. It was previously sending pictures based on the last picture taken/used.

In this new iteration it solves the issue where people in different games could possibly receive images that their friends didn't take, but someone randomly took for another game